requirements.txt
flask
flask-cors
praw
pandas
schedule
python-dotenv
pytz